+++ packages/workspace-fusion-modules/src/lib/modules/fusionBookmarksModule/utils/bookmarksApi.ts
@@ -1,0 +1,6 @@
+import { getBookmarkById } from '../api/getBookmarkById';
+
+/** Api for using bookmarks */
+export const bookmarksApi = {
+	getBookmarkById,
+};
